improvement(Code refactor): Move feed version extraction into FeedVersionSummary and created separat
