+++ src/main/java/com/conveyal/datatools/manager/models/FeedVersionSummary.java
@@ -2,23 +2,28 @@ package com.conveyal.datatools.manager.models;
 
 import com.conveyal.datatools.editor.utils.JacksonSerializers;
 import com.conveyal.datatools.manager.gtfsplus.GtfsPlusValidation;
-import com.conveyal.gtfs.graphql.fetchers.ErrorCountFetcher;
+import com.conveyal.datatools.manager.gtfsplus.ValidationIssue;
 import com.conveyal.gtfs.validator.ValidationResult;
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.bson.Document;
 
 import java.io.Serializable;
 import java.time.LocalDate;
-import java.util.ArrayList;
+import java.time.format.DateTimeFormatter;
 import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Includes summary data (a subset of fields) for a feed version.
  */
 public class FeedVersionSummary extends Model implements Serializable {
     private static final long serialVersionUID = 1L;
+    private static final ObjectMapper mapper = new ObjectMapper();
+    private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyyMMdd");
 
     public FeedRetrievalMethod retrievalMethod;
     public int version;
@@ -39,7 +44,6 @@ public class FeedVersionSummary extends Model implements Serializable {
     public int publishedFeedVersionErrorCount;
     public LocalDate publishedFeedVersionStartDate;
     public LocalDate publishedFeedVersionEndDate;
-    public List<ErrorCountFetcher.ErrorCount> errorCounts = new ArrayList<>();
 
     public PartialValidationSummary getValidationSummary() {
         if (validationSummary == null) {
@@ -51,6 +55,33 @@ public class FeedVersionSummary extends Model implements Serializable {
     /** Empty constructor for serialization */
     public FeedVersionSummary() {
         // Do nothing
+    }
+
+    public FeedVersionSummary(
+        String feedVersionKey,
+        boolean hasChildValidationResultDocument,
+        Document feedVersionDocument
+    ) {
+        id = feedVersionDocument.getString(feedVersionKey);
+        processedByExternalPublisher = feedVersionDocument.getDate("processedByExternalPublisher");
+        sentToExternalPublisher = feedVersionDocument.getDate("sentToExternalPublisher");
+        gtfsPlusValidation = getGtfsPlusValidation(feedVersionDocument);
+        namespace = feedVersionDocument.getString("namespace");
+        validationResult = getValidationResult(hasChildValidationResultDocument, feedVersionDocument);
+
+        // The feed source's published version id.
+        publishedVersionId = feedVersionDocument.getString("publishedVersionId");
+
+        // The feed source's published feed version. Feed source's publishedVersionId mapped to feed version's namespace.
+        publishedFeedVersionErrorCount = feedVersionDocument.get("publishedFeedVersionErrorCount") != null
+            ? feedVersionDocument.getInteger("publishedFeedVersionErrorCount")
+            : 0;
+        publishedFeedVersionStartDate = feedVersionDocument.get("publishedFeedVersionStartDate") != null
+            ? getDateFromString(feedVersionDocument.getString("publishedFeedVersionStartDate"))
+            : null;
+        publishedFeedVersionEndDate = feedVersionDocument.get("publishedFeedVersionEndDate") != null
+            ? getDateFromString(feedVersionDocument.getString("publishedFeedVersionEndDate"))
+            : null;
     }
 
     /**
@@ -74,5 +105,96 @@ public class FeedVersionSummary extends Model implements Serializable {
                 this.endDate = validationResult.lastCalendarDate;
             }
         }
+    }
+
+    /**
+     * Build GtfsPlusValidation object from feed version document.
+     */
+    private static GtfsPlusValidation getGtfsPlusValidation(Document feedVersionDocument) {
+        Document gtfsPlusValidationDocument = getDocumentChild(feedVersionDocument, "gtfsPlusValidation");
+        if (gtfsPlusValidationDocument == null) {
+            return null;
+        }
+        List<ValidationIssue> issues = null;
+        if (gtfsPlusValidationDocument.containsKey("issues") && gtfsPlusValidationDocument.get("issues") != null) {
+            List<Document> issueDocs = gtfsPlusValidationDocument.getList("issues", Document.class);
+            issues = issueDocs
+                .stream()
+                .map(doc -> mapper.convertValue(doc, ValidationIssue.class))
+                .collect(Collectors.toList());
+        }
+        boolean published = Boolean.TRUE.equals(gtfsPlusValidationDocument.getBoolean("published"));
+        return new GtfsPlusValidation(published, issues);
+    }
+
+    /**
+     * Build validation result from feed version document.
+     */
+    private static ValidationResult getValidationResult(boolean hasChildValidationResultDocument, Document feedVersionDocument) {
+        ValidationResult validationResult = new ValidationResult();
+        validationResult.errorCount = getValidationResultErrorCount(hasChildValidationResultDocument, feedVersionDocument);
+        validationResult.firstCalendarDate = getValidationResultDate(hasChildValidationResultDocument, feedVersionDocument, "firstCalendarDate");
+        validationResult.lastCalendarDate = getValidationResultDate(hasChildValidationResultDocument, feedVersionDocument, "lastCalendarDate");
+        return validationResult;
+    }
+
+    /**
+     * Convert String date (if not null) into LocalDate.
+     */
+    private static LocalDate getDateFromString(String date) {
+        return (date == null) ? null : LocalDate.parse(date, formatter);
+    }
+
+    /**
+     * Extract child document matching provided name.
+     */
+    private static Document getDocumentChild(Document document, String name) {
+        return (Document) document.get(name);
+    }
+
+    /**
+     * Extract date value from parent document or child validation result document.
+     */
+    private static LocalDate getValidationResultDate(
+        boolean hasChildValidationResultDocument,
+        Document feedVersionDocument,
+        String key
+    ) {
+        return (hasChildValidationResultDocument)
+            ? getDateFieldFromDocument(feedVersionDocument, key)
+            : getDateFromString(feedVersionDocument.getString(key));
+    }
+
+    /**
+     * Extract date value from validation result document.
+     */
+    private static LocalDate getDateFieldFromDocument(Document document, String dateKey) {
+        Document validationResult = getDocumentChild(document, "validationResult");
+        return (validationResult != null)
+            ? getDateFromString(validationResult.getString(dateKey))
+            : null;
+    }
+
+    /**
+     * Extract the error count from the parent document or child validation result document. If the error count is not
+     * available, return -1.
+     */
+    private static int getValidationResultErrorCount(boolean hasChildValidationResultDocument, Document feedVersionDocument) {
+        int errorCount;
+        try {
+            errorCount = (hasChildValidationResultDocument)
+                ? getErrorCount(feedVersionDocument)
+                : feedVersionDocument.getInteger("errorCount");
+        } catch (NullPointerException e) {
+            errorCount = -1;
+        }
+        return errorCount;
+    }
+
+    /**
+     * Get the child validation result document and extract the error count from this.
+     */
+    private static int getErrorCount(Document document) {
+        return getDocumentChild(document, "validationResult").getInteger("errorCount");
     }
 }
