+++ src/test/java/com/conveyal/datatools/manager/controllers/api/FeedSourceControllerTest.java
@@ -44,7 +44,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class FeedSourceControllerTest extends DatatoolsTest {
     private static Project project = null;
@@ -55,21 +54,14 @@ public class FeedSourceControllerTest extends DatatoolsTest {
     private static FeedSource feedSourceWithInvalidLabels = null;
     private static Label publicLabel = null;
     private static Label adminOnlyLabel = null;
-    private static Label feedSourceWithLatestDeploymentAdminOnlyLabel = null;
-    private static Label feedSourceWithPinnedDeploymentAdminOnlyLabel = null;
-    private static Note feedSourceWithLatestDeploymentAdminOnlyNote = null;
-    private static Note feedSourceWithPinnedDeploymentAdminOnlyNote = null;
     private static Project projectWithLatestDeployment = null;
     private static FeedSource feedSourceWithLatestDeploymentFeedVersion = null;
     private static FeedVersion feedVersionFromLatestDeployment = null;
     private static FeedVersion feedVersionPublishedFromLatestDeployment = null;
-    private static Deployment deploymentLatest = null;
-    private static Deployment deploymentSuperseded = null;
 
     private static Project projectWithPinnedDeployment = null;
     private static FeedSource feedSourceWithPinnedDeploymentFeedVersion = null;
     private static FeedVersion feedVersionFromPinnedDeployment = null;
-    private static Deployment deploymentPinned = null;
 
     @BeforeAll
     public static void setUp() throws IOException {
@@ -77,16 +69,8 @@ public class FeedSourceControllerTest extends DatatoolsTest {
         Auth0Connection.setAuthDisabled(true);
         ProcessSingleFeedJob.ENABLE_ADDITIONAL_VALIDATION = false;
 
-        project = new Project();
-        project.name = "ProjectOne";
-        project.autoFetchFeeds = true;
-        Persistence.projects.create(project);
-
-        projectToBeDeleted = new Project();
-        projectToBeDeleted.name = "ProjectTwo";
-        projectToBeDeleted.autoFetchFeeds = false;
-        Persistence.projects.create(projectToBeDeleted);
-
+        project = createProject("ProjectOne", true);
+        projectToBeDeleted = createProject("ProjectTwo", false);
 
         feedSourceWithUrl = createFeedSource("FeedSourceOne", new URL("http://www.feedsource.com"), project);
         feedSourceWithNoUrl = createFeedSource("FeedSourceTwo", null, project);
@@ -99,7 +83,6 @@ public class FeedSourceControllerTest extends DatatoolsTest {
 
         setUpFeedVersionFromLatestDeployment();
         setUpFeedVersionFromPinnedDeployment();
-
     }
 
     /**
@@ -111,8 +94,10 @@ public class FeedSourceControllerTest extends DatatoolsTest {
         projectWithLatestDeployment.organizationId = "project-with-latest-deployment-org-id";
         Persistence.projects.create(projectWithLatestDeployment);
 
-        feedSourceWithLatestDeploymentAdminOnlyLabel = createLabel("label-id-latest-deployment", "Admin Only Label", projectWithLatestDeployment.id);
-        feedSourceWithLatestDeploymentAdminOnlyNote = createNote("note-id-latest-deployment", "A test note");
+        Label feedSourceWithLatestDeploymentAdminOnlyLabel = createLabel(
+            "label-id-latest-deployment", "Admin Only Label", projectWithLatestDeployment.id
+        );
+        Note feedSourceWithLatestDeploymentAdminOnlyNote = createNote("note-id-latest-deployment", "A test note");
 
         feedSourceWithLatestDeploymentFeedVersion = createFeedSource(
             "feed-source-with-latest-deployment-feed-version",
@@ -135,10 +120,14 @@ public class FeedSourceControllerTest extends DatatoolsTest {
             null
         );
 
-        FeedVersion feedVersionFromGtfsZip = createFeedVersionFromGtfsZip(feedSourceWithLatestDeploymentFeedVersion, "bart_old.zip");
+        FeedVersion feedVersionFromGtfsZip = createFeedVersionFromGtfsZip(
+            feedSourceWithLatestDeploymentFeedVersion,
+            "bart_old.zip"
+        );
         // Update the feed version namespace to match that created from the import.
         feedVersionFromLatestDeployment.namespace = feedVersionFromGtfsZip.namespace;
 
+        // Remove the imported feed version so it does not conflict with the latest deployment feed version.
         Persistence.feedVersions.removeById(feedVersionFromGtfsZip.id);
         Persistence.feedVersions.replace(feedVersionFromLatestDeployment.id, feedVersionFromLatestDeployment);
 
@@ -150,13 +139,13 @@ public class FeedSourceControllerTest extends DatatoolsTest {
             LocalDate.of(2022, Month.NOVEMBER, 3),
             feedSourceWithLatestDeploymentFeedVersion.publishedVersionId
         );
-        deploymentSuperseded = createDeployment(
+        createDeployment(
             "deployment-superseded",
             projectWithLatestDeployment,
             feedVersionFromLatestDeployment.id,
             deployedSuperseded
         );
-        deploymentLatest = createDeployment(
+        createDeployment(
             "deployment-latest",
             projectWithLatestDeployment,
             feedVersionFromLatestDeployment.id,
@@ -173,8 +162,8 @@ public class FeedSourceControllerTest extends DatatoolsTest {
         projectWithPinnedDeployment.organizationId = "project-with-pinned-deployment-org-id";
         Persistence.projects.create(projectWithPinnedDeployment);
 
-        feedSourceWithPinnedDeploymentAdminOnlyLabel = createLabel("label-id-pinned-deployment", "Admin Only Label", projectWithPinnedDeployment.id);
-        feedSourceWithPinnedDeploymentAdminOnlyNote = createNote("note-id-pinned-deployment", "A test note");
+        Label feedSourceWithPinnedDeploymentAdminOnlyLabel = createLabel("label-id-pinned-deployment", "Admin Only Label", projectWithPinnedDeployment.id);
+        Note feedSourceWithPinnedDeploymentAdminOnlyNote = createNote("note-id-pinned-deployment", "A test note");
 
         feedSourceWithPinnedDeploymentFeedVersion = createFeedSource(
             "feed-source-with-pinned-deployment-feed-version",
@@ -190,7 +179,7 @@ public class FeedSourceControllerTest extends DatatoolsTest {
             feedSourceWithPinnedDeploymentFeedVersion.id,
             LocalDate.of(2022, Month.NOVEMBER, 2)
         );
-        deploymentPinned = createDeployment(
+        Deployment deploymentPinned = createDeployment(
             "deployment-pinned",
             projectWithPinnedDeployment,
             feedVersionFromPinnedDeployment.id,
@@ -204,74 +193,30 @@ public class FeedSourceControllerTest extends DatatoolsTest {
     public static void tearDown() {
         Auth0Connection.setAuthDisabled(Auth0Connection.getDefaultAuthDisabled());
         if (project != null) {
-            Persistence.projects.removeById(project.id);
+            project.delete();
         }
         if (projectToBeDeleted != null) {
-            Persistence.projects.removeById(projectToBeDeleted.id);
-        }
-        if (feedSourceWithUrl != null) {
-            Persistence.feedSources.removeById(feedSourceWithUrl.id);
-        }
-        if (feedSourceWithNoUrl != null) {
-            Persistence.feedSources.removeById(feedSourceWithNoUrl.id);
-        }
-        if (publicLabel != null) {
-            Persistence.labels.removeById(publicLabel.id);
-        }
-        if (adminOnlyLabel != null) {
-            Persistence.labels.removeById(adminOnlyLabel.id);
+            projectToBeDeleted.delete();
         }
         tearDownDeployedFeedVersion();
         ProcessSingleFeedJob.ENABLE_ADDITIONAL_VALIDATION = true;
     }
 
     /**
-     * These entities are removed separately so that if the need arises they can be kept.
+     * These projects are removed separately so that if the need arises they can be kept.
      * This would then allow the Mongo queries defined in FeedSource#getFeedVersionFromLatestDeployment and
      * FeedSource#getFeedVersionFromPinnedDeployment to be tested.
      */
     private static void tearDownDeployedFeedVersion() {
-        if (feedVersionFromLatestDeployment != null) {
-            feedVersionFromLatestDeployment.delete();
-        }
         if (projectWithPinnedDeployment != null) {
             projectWithPinnedDeployment.delete();
         }
         if (projectWithLatestDeployment != null) {
-            Persistence.projects.removeById(projectWithLatestDeployment.id);
+            projectWithLatestDeployment.delete();
         }
-        if (feedSourceWithLatestDeploymentFeedVersion != null) {
-            Persistence.feedSources.removeById(feedSourceWithLatestDeploymentFeedVersion.id);
-        }
-        if (feedSourceWithPinnedDeploymentFeedVersion != null) {
-            Persistence.feedSources.removeById(feedSourceWithPinnedDeploymentFeedVersion.id);
-        }
+        // Orphan feed version must be explicitly deleted.
         if (feedVersionPublishedFromLatestDeployment != null) {
             Persistence.feedVersions.removeById(feedVersionPublishedFromLatestDeployment.id);
-        }
-        if (feedVersionFromPinnedDeployment != null) {
-            Persistence.feedVersions.removeById(feedVersionFromPinnedDeployment.id);
-        }
-        if (deploymentPinned != null) {
-            Persistence.deployments.removeById(deploymentPinned.id);
-        }
-        if (deploymentLatest != null) {
-            Persistence.deployments.removeById(deploymentLatest.id);
-        }
-        if (deploymentSuperseded != null) {
-            Persistence.deployments.removeById(deploymentSuperseded.id);
-        }
-        if (feedSourceWithPinnedDeploymentAdminOnlyLabel != null) {
-            Persistence.labels.removeById(feedSourceWithPinnedDeploymentAdminOnlyLabel.id);
-        }
-        if (feedSourceWithLatestDeploymentAdminOnlyLabel != null) {
-            Persistence.labels.removeById(feedSourceWithLatestDeploymentAdminOnlyLabel.id);
-        }
-        if (feedSourceWithPinnedDeploymentAdminOnlyNote != null) {
-            Persistence.notes.removeById(feedSourceWithPinnedDeploymentAdminOnlyNote.id);
-        }
-        if (feedSourceWithLatestDeploymentAdminOnlyNote != null) {
-            Persistence.notes.removeById(feedSourceWithLatestDeploymentAdminOnlyNote.id);
         }
     }
 
@@ -368,8 +313,8 @@ public class FeedSourceControllerTest extends DatatoolsTest {
         // Test that they are assigned properly
         assertEquals(2, labelCountForFeed(feedSourceWithLabels.id));
         // Test that project shows only correct labels based on user auth
-        assertEquals(2, labelCountforProject(feedSourceWithLabels.projectId, true));
-        assertEquals(1, labelCountforProject(feedSourceWithLabels.projectId, false));
+        assertEquals(2, labelCountForProject(feedSourceWithLabels.projectId, true));
+        assertEquals(1, labelCountForProject(feedSourceWithLabels.projectId, false));
 
         // Test that feed source shows only correct labels based on user auth
         List<String> labelsSeenByAdmin = FeedSourceController.cleanFeedSourceForNonAdmins(feedSourceWithLabels, true).labelIds;
@@ -385,7 +330,7 @@ public class FeedSourceControllerTest extends DatatoolsTest {
         );
         assertEquals(OK_200, deleteSecondLabelResponse.status);
         assertEquals(1, labelCountForFeed(feedSourceWithLabels.id));
-        assertEquals(1, labelCountforProject(feedSourceWithLabels.projectId, true));
+        assertEquals(1, labelCountForProject(feedSourceWithLabels.projectId, true));
 
         // Test that labels are removed when deleting project
         assertEquals(1, Persistence.labels.getFiltered(eq("projectId", projectToBeDeleted.id)).size());
@@ -483,6 +428,14 @@ public class FeedSourceControllerTest extends DatatoolsTest {
         assertEquals(feedVersionFromPinnedDeployment.gtfsPlusValidation.published, feedSourceSummaries.get(0).latestGtfsPlusValidation.published);
     }
 
+    private static Project createProject(String name, boolean autoFetchFeeds) {
+        Project project = new Project();
+        project.name = name;
+        project.autoFetchFeeds = autoFetchFeeds;
+        Persistence.projects.create(project);
+        return project;
+    }
+
     private static FeedSource createFeedSource(String name, URL url, Project project) {
         return createFeedSource(null, name, url, project, false);
     }
@@ -541,10 +494,6 @@ public class FeedSourceControllerTest extends DatatoolsTest {
      */
     private static FeedVersion createFeedVersion(String id, String feedSourceId, LocalDate endDate) {
         return createFeedVersion(id, feedSourceId, null, endDate, null);
-    }
-
-    private static FeedVersion createFeedVersion(String id, String feedSourceId, LocalDate endDate, String namespace) {
-        return createFeedVersion(id, feedSourceId, null, endDate, namespace);
     }
 
     /**
@@ -618,7 +567,7 @@ public class FeedSourceControllerTest extends DatatoolsTest {
     /**
      * Provide the label count for a given project
      */
-    private int labelCountforProject(String projectId, boolean isAdmin) {
+    private int labelCountForProject(String projectId, boolean isAdmin) {
         return Persistence.projects.getById(projectId).retrieveProjectLabels(isAdmin).size();
     }
 }
